requirements.txt
requests
fastapi
pydantic
python-docx
python-jose
uvicorn